서버의 pch헤더가 포함되는 것 같아서 리팩토링+서버에 XMFLOAT3도입+3차원 이동준비
PIP최상위 폴더안에 Common 폴더 생성해서 안에 Packet, Vector3, PacketStream 옮김 근데 Vector3가 클라에 있어서 이걸 저기로 옮기는 방안 검토가 필요함(대원, 찬진~ 검토좀)
그리고 메인 캐릭터 이동 할때 이제는 방향을 벡터값으로 넣어줘야함 일단은
월드 좌표기준으로 방향을 넣어줘야하는데 어케 넣을지 몰라서 일단 비워둠(클라 "해줘")
서버쪽에서 벡터 도입

+++ Chess_Client/Chess_Client/OtherPlayer.h
@@ -20,7 +20,7 @@ public:
 	int64_t GetID() const { return _id; }
 	void SetName(const std::string& name) { _name = name; }
 	std::string GetName() const { return _name; }
-	void SetPos(int x = 0, int y = 0) { _NowX = x; _NowY = y; }
+	void SetPos(float x = 0, float y = 0) { _NowX = x; _NowY = y; }
 
 private:
